requirements.txt
python-dotenv==1.0.1
python-multipart==0.0.20
pydantic==2.10.4
psycopg2-binary==2.9.10
fastapi==0.115.6
uvicorn==0.34.0
mypy==1.14.0
bcrypt==4.0.1
sqlalchemy==2.0.36